requirements.txt
djangorestframework~=3.16.1
markdown
django-filter
requests
Django~=5.2.5
django-oauth-toolkit
django-cors-headers
python-magic
python-magic-bin
djangorestframework-simplejwt
croniter